templates/Web/_composition/Flask/Feature.Flask.Azure.Cosmos.SQL/backend/requirements_postaction.txt
//{[{
azure-cosmos==4.2.0
python-dotenv==0.15.0
//}]}
flask==1.1.2